use auth for delete function

+++ server/routes/consumables.ts
@@ -56,17 +56,32 @@ router.post('/', async (req, res) => {
   }
 })
 
-router.delete('/:id', async (req: JwtRequest, res, next) => {
-  // if (!req.auth?.sub) {
-  //   res.sendStatus(StatusCodes.UNAUTHORIZED)
-  //   return
-  // }
+router.delete('/:id', checkJwt, async (req: JwtRequest, res, next) => {
+  console.log('Auth:', req.auth)
+  if (!req.auth?.sub) {
+    res.sendStatus(StatusCodes.UNAUTHORIZED)
+    return
+  }
+  const id = Number(req.params.id)
+  const auth0Id = req.auth ? req.auth.sub : undefined
+  if (!id) {
+    console.error('Invalid id')
+    return res.status(400).send('Bad request')
+  }
+
+  if (!auth0Id) {
+    console.error('No auth0Id')
+    return res.status(401).send('Unauthorized')
+  }
+
   try {
-    const id = Number(req.params.id)
     await db.deleteConsumable(id)
-  } catch (e) {
-    next(e)
+    res
+      .setHeader('Location', `${req.baseUrl}/${id}`)
+      .sendStatus(StatusCodes.CREATED)
+  } catch (err) {
     res.status(500).json({ message: 'It is not deleted. Try again' })
+    next(err)
   }
 })
 export default router
